closes DRAG-40
This allows Wink/Arduino projects to specify ports/board correctly for wink only the port needs to be configured.

LoadedProject now has a getMetaData method to make the code more terse, the core arduino packge has been update to use this method which correctly returns the board and port that a project is using. 

Migrations have been add to Ringo and Wink to add in the board data, Ringo requires setting a processor which is not supported at the moment so the project type is currently disabled see DRAG-53 for details