Moving game object point functions to use point class to match the physics category will also create boolean/numeric blocks to match as well.
See addtion of createDropdownField function this will take a list of fields that are writable and list of fields that are read only and create the correct array of arrays to use for the blocks. This will add some computation on load but is always possible to move to a cmd line utility that spits out arrays to copy and paste if it becomes too intensive.