Added in Icons to store all icon versions of blocks
The hierarchy in this folder is a copy of the normal blocks, generators

These blocks are allowed to sue Node.JS unlike normal blocks there are
some edge cases with using Node in this manner __dirname in particular
for example will be the html files location not the JavaScript file
location.

Wink still needs to have a new Exploring Wink Bot created and blocks
converted to use icons.

Text under icons will be removed in theory