electron-logger console/file logging utility does 90% of what is needed to make logging cleaner then it currently is (varying log levels)
Projects are now created using an archive file format. At the moment
this means that the actual project is just stored in a temp folder and
then zipped up to ~/${USER}/Documents/DragonDropProjects all changes
are saved to the copy in the temp folder.

Recent files now searches for both .digiblocks and .drop files with
regards to local files.