Save As has been refactored to use ipc to communicate the desire to save a copy to the open project.
First a new project is created to make the base files exist.

This project is passed through ipc to the render thread which reuses
the save method passing in the new project with the current projects
data. This means the update to the .drop happens automatically

The project is then sent back to the main process to be displayed.